refactor: move tooltip to types/xy_chart/tooltip
